requirements.txt
flask
gtts
pymupdf
pytesseract
pillow
gunicorn
